Renames module for clarity
